Restructured File structure
