+++ src/Pages/Archives/Components/archives.jsx
@@ -1,6 +1,6 @@
 import {Component} from 'react'
 
-import HalfTitle from '../../_Common/Components/halfTitle';
+import HalfTitle from '../../../components/HalfTitle/HalfTitle';
 
 import FileGallery from './archivesFileGallery';
