+++ src/Pages/Leadership/Components/leadership.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from 'react'
 
-import HalfTitle from '../../_Common/Components/halfTitle';
+import HalfTitle from '../../../components/HalfTitle/HalfTitle';
 
 import LeadershipGallery from './leadershipGallery';
 import LeadershipIntro from './leadershipIntro';
